// libs/cms-ui-library/src/lib/cms-api/v1/models/CourtCase.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CourtCase = {
    id: string;
    caseNumber: string;
    location: string;
    plaintiff: string;
    defendant: string;
    status: string;
    type: string;
    outcome: string;
    lawyerId: string;
    userId: string;
    dateCreated: string;
};